fix(discord): extract revoke wording to discord.js-free module so e2e smoke can load it
PR #198 introduced `e2e/tests/qurl-send-revoke.smoke.test.ts` which
imported the bot's `commands.js` directly to assert on
`renderRevokeMsg`'s output. After merge the post-deploy smoke step
fails to load with `Cannot find module 'discord.js' from
'../apps/discord/src/commands.js'` — the e2e CI runner only
installs `e2e/` deps, so `apps/discord/node_modules/discord.js`
isn't present.

Extracted the pure-string wording layer (`renderRevokeContent`,
`buildRevokeHeader`, `REVOKE_TRUNC_LIMIT`,
`REVOKE_CONTENT_SAFE_MAX`) into a new
`apps/discord/src/revoke-render.js` whose only require is
`./utils/sanitize` (also discord.js-free). The smoke test now
imports from `revoke-render` and loads cleanly with no
`discord.js`/store/config/logger transitives needed.

`commands.js`'s `renderRevokeMsg` is now a thin wrapper that calls
`renderRevokeContent` for content + `attachmentText`, then builds
the discord.js Show All `ActionRowBuilder` from `needsExpand` /
`showAll` fields. Existing bot-side unit tests in
`qurl-send-back-half.test.js` continue to pass against the
wrapper's exported `_test.renderRevokeMsg` shape.

Verified locally:
- apps/discord: 934 tests + eslint --max-warnings 0 green
- e2e: smoke test 6/6 pass without discord.js in node_modules

Auto-revert on smoke failure was skipped on the merge run because
the smoke step exited 1 before the revert gate's signal could fire
— the bot is deployed in sandbox; this PR only restores the post-
deploy wording-drift gate.
