+++ apps/discord/src/commands.js
@@ -2554,8 +2554,13 @@ async function handleRevoke(interaction, apiKey) {
   }
 }
 
-// Shared truncation limit for both send + revoke recipient lists.
-const REVOKE_TRUNC_LIMIT = 5;
+// Pure rendering / wording logic lives in `./revoke-render` so the
+// e2e smoke test can require it without pulling in `discord.js`.
+const {
+  REVOKE_TRUNC_LIMIT,
+  buildRevokeHeader,
+  renderRevokeContent,
+} = require('./revoke-render');
 
 // Builds the editReply payload from a `renderRevokeMsg` result. When
 // the rendered names list overflowed the Discord content cap,
@@ -2572,76 +2577,21 @@ function revokeReplyPayload(rendered) {
   return payload;
 }
 
-// Discord message content cap is 2000 chars. Leave headroom for the
-// header + "Revoked for: " prefix; force truncation in renderRevokeMsg
-// even when showAll=true if the full names list would push the total
-// over this. Without it, a Show All on ~80+ recipients would exceed
-// Discord's limit and the editReply would error.
-const REVOKE_CONTENT_SAFE_MAX = 1900;
-
-// Single source of truth for the "Revoked X/Y users[.]" header +
-// already-opened note. Used by both the inline-button path (via
-// renderRevokeMsg) and the slash-command /qurl revoke handler so a
-// future wording change lands in one place.
-function buildRevokeHeader(success, total) {
-  const note = total > 0 ? ' Note: already-opened links cannot be revoked.' : '';
-  return `Revoked ${success}/${total} user${total !== 1 ? 's' : ''}.${note}`;
-}
-
-// Builds the post-revoke confirmation message + Show All toggle row.
-// User-centric: `success`/`total` are unique-recipient counts; `names`
-// is the strict-success list. Returns `attachmentText` (newline-joined
-// full list) when the inline rendering would exceed Discord's 2000-char
-// content cap — caller wraps in an AttachmentBuilder. In attachment
-// mode the Show All button is suppressed since the file IS the full
-// list.
-//
-// `success` defaults to `names.length`. Pass it explicitly when the
-// authoritative count (e.g. DDB strict-success) may exceed the names
-// the caller could resolve — header reflects truth, names list
-// reflects what's renderable.
+// Wraps `renderRevokeContent` (pure data) and adds the discord.js
+// ActionRowBuilder/ButtonBuilder for the Show All / Show Less toggle.
+// All wording assertions live against `renderRevokeContent` directly
+// (see `apps/discord/src/revoke-render.js` + the e2e smoke).
 function renderRevokeMsg(sendId, names, total, showAll, success = names.length) {
-  let content = buildRevokeHeader(success, total);
-  let attachmentText = null;
-
-  if (names.length === 0) {
-    return { content, needsExpand: false, row: null, attachmentText };
-  }
-
-  // `names` are plain (sanitizeDisplayNamePlain at the call site) so
-  // they can land verbatim in the .txt attachment. Message content
-  // needs markdown escape per name to defuse `*phish*` / `[t](url)`
-  // injection — render-context split.
-  const escapedNames = names.map(escapeDiscordMarkdown);
-  const fullLine = `\nRevoked for: ${escapedNames.join(', ')}`;
-  const fullFits = content.length + fullLine.length <= REVOKE_CONTENT_SAFE_MAX;
-
-  if (!fullFits) {
-    // Full list won't fit inline → emit as a file attachment. Inline
-    // shows the first REVOKE_TRUNC_LIMIT names + "(see attached)" pointer.
-    const preview = escapedNames.slice(0, REVOKE_TRUNC_LIMIT).join(', ');
-    content += `\nRevoked for: ${preview} +${names.length - REVOKE_TRUNC_LIMIT} more (see attached)`;
-    attachmentText = names.join('\n');
-    return { content, needsExpand: false, row: null, attachmentText };
-  }
-
-  // Full list fits inline — current Show All / Show Less behavior.
-  if (showAll || names.length <= REVOKE_TRUNC_LIMIT) {
-    content += fullLine;
-  } else {
-    content += `\nRevoked for: ${escapedNames.slice(0, REVOKE_TRUNC_LIMIT).join(', ')} +${names.length - REVOKE_TRUNC_LIMIT} more`;
-  }
-
-  const needsExpand = names.length > REVOKE_TRUNC_LIMIT;
-  const row = needsExpand
+  const data = renderRevokeContent(sendId, names, total, showAll, success);
+  const row = data.needsExpand
     ? new ActionRowBuilder().addComponents(
       new ButtonBuilder()
-        .setCustomId(`qurl_revoke_expand_${sendId}`)
-        .setLabel(showAll ? 'Show Less' : 'Show All')
+        .setCustomId(`qurl_revoke_expand_${data.sendId}`)
+        .setLabel(data.showAll ? 'Show Less' : 'Show All')
         .setStyle(ButtonStyle.Secondary),
     )
     : null;
-  return { content, needsExpand, row, attachmentText };
+  return { content: data.content, needsExpand: data.needsExpand, row, attachmentText: data.attachmentText };
 }
 
 async function revokeAllLinks(sendId, senderDiscordId, apiKey) {
